initial commit 'stellar'.
This may be a working title (again).

Stellar is going to be for making constellations of builds with
repeatr.  This is not unlike the description of the 'heft' project,
but A) I'm tired of *that* working title, and B) it's time for
a clean slate again so that we can use the redux version of the
go-timeless-api code (and also, leave aside the scripting bits
until much later in the design process, because maintaining those
bindings while anything else is in flux is really makeworky).

(Heft will probably be merge-ignored in eventually; a lot of code
*will* come from there, e.g. the commission graph traversal logic is
already thoroughly not-wrong; but it's easier to whitelist things
piecemeal while simultaneously managing the api types port.)

Signed-off-by: Eric Myhre <hash@exultant.us>