Module evaluation, for real, using repeatr.
I'm a tad shocked that this A) works and B) in less than 500 lines of code.

(Calling it "500 lines" is ignoring a *lot* of work in libraries and API
basics leading up to this point, of course, but still.)

Caveats:

- submodule evaluation isn't implemented yet.  (Going to start a new fixture
  to cover that before proceeding.)
- the warehouse handling is bunkum and hardcoded (and antibonus, also in-tree
  until we git around to improving it).
- populating the warehouse with the ware needed for the demo is not included.
  (It'll just work if you happen to have it cached.)
- ware caching isn't isolated to the tests, speaking of which.

... so in general: this is quite fast and loose with everything that you can
possibly consider hidden behind the repeatr API layer.

And actually, I'm kind of okay with that.  (For now, at least.)  It's a cool
proof of concept.  We'll rachet the testing rigour back up later.

Signed-off-by: Eric Myhre <hash@exultant.us>