Hacky first pass on forwarding repeatr logs.
This gets us Interesting Things to look at again; but is very messy.
We need to introduce output control handles throughout stellar
presently in order to make this move forward gracefully.
(Which is... going to be a bit of a fuss, since I've already started
using the placeholder output as test fixtures...)

Signed-off-by: Eric Myhre <hash@exultant.us>