Proper use of wareSourcing.
The ".timeless/warehouse" default was previously hardcoded in the
repeatr API client; now it's moved up to us here in stellar.main.
We'll continue to push it further (to env vars, etc) later.

And importantly, it's now only applied to "tar"-type wares!
Which gets us a step closer to git being handled correctly.

Signed-off-by: Eric Myhre <hash@exultant.us>