Warehouses can now be recommended by 'mirror.tl'.
This fixes one of the last major core usability hurdles.  We can now
being to assemble a catalog repo full of release info, as well as
pointer on which warehouses to get stuff from... and between the
timeless stack core tool binaries and that info, anyone should be
able to build (both in theory and now in practice)... anything.
Anywhere.

Automated tests are sparse here, but for now you can exercise this
manually by setting `RIO_CACHE=/tmp/some-$RANDOM-path` in your env to
make sure you have no usable cache, and then rely on the mirror.tl
files in the catalog paths to make your operations workable.
Previously, that would leave you in trouble; now it should fly fine.

Signed-off-by: Eric Myhre <hash@exultant.us>