Begin to upgrade basic printfs toward structured logging.
Immediately derailed by attempts to use go templates.  It's clearly specified
that 'range' operations will traverse maps in sorted order if keys are
comparable, and if those structs that are keys in the pins map aren't
comparable then I eat my hat in protest.  And yet, our tests on output messages
now sporadically fail as if that template is not sorting as it claims.

So this is going into the merge-ignore dustbin almost immediately.
Parts may be revived later -- having some flags for controlling output format
is certainly a must -- but it'll be a different approach than this.

(We also probably ought to skip this templating nonsense and proceed directly
to well-typed "message" formats like those repeatr and the other timeless
stack components use here.  I suspect 'stellar' might be the practical top of
most process trees -- invoked by the user -- but we should still prepare to be
a good API-driven citizen.)

Signed-off-by: Eric Myhre <hash@exultant.us>