Support for "ingest:literal:tar:f00bar" imports.
Which is *extremely limited*, note the comment about WareSourcing.

At present it's probably a much better idea to just make a catalog with the
thing that you want to ingest, because this gives us a clear place to put any
warehouse addresses: in the mirrors.tl metadata file in the catalog.

Signed-off-by: Eric Myhre <hash@exultant.us>