New package with the beginnings of catalog linter.
Writing code for things which is in practice closely bound to the filesystem
is always a special kind of fun.

Perhaps someday we'll want to make this more pluggable, but at present that
would be abstraction without a purpose nor a proving ground.

It also sort of seems like the walkfunc will only carry us so far -- once we
find a dir that's got *any* of our files in it, we need to load *all* of
them to start verification of any cross-referencing parts.

Signed-off-by: Eric Myhre <hash@exultant.us>