'stellar catalog lint' now canonicalizes.
It does this only if the '--rewrite' flag is given.

This behavior just blazes ahead in all cases; it will not log how many files
were actually *changed* as a result of this.  PRs welcome.

(Since at present we're keeping the catalog files in a git repo, git diff
suffices for telling us if something changed.  This should be regarded as a
stopgap though; PRs really *are* welcome to add independent accounting here.)

Signed-off-by: Eric Myhre <hash@exultant.us>