Revamp layout.Landmarks for workspaces.
Also, allow the module.tl file to be a sibling of the dot-timeless dir.

The workspaces functionality will be coming later -- it's going to be a LOT.

Note the naming of Landmarks.ModuleCatalogRoot.  There is no comparable
'Landmarks.WorkspaceCatalogRoot'.  Why not?
Module catalogs are planned to move towards being a vendored subset of catalog
data needed to build the module.
The workspace scope is intended to be where you will keep the full public
catalog info trees (and fetch updates to the catalog info).
The workspace scope also isn't going to be limited to *one catalog*; you'll be
able to configure overlays of multiple catalogs as desired.  (One anticipated
common pattern will be a "main" catalog plus a "staging" catalog which contains
releases that you've drafted but not pushed publicly, for example).

Why are we promoting the module.tl file to the project root?
Because it's important: it's the only thing that's the *public* API to
timelessness of a project.
If there's module generation scripts, templates, update selectors... those are
in the dot-timeless dir; and this is rightful, because those are though they
are also valuable, they are *not* a guaranteed part of the core standard and
intentionally excluded from public API that you need to know about to evaluate
a module.
Same goes for the .timeless/catalog dir: though it is fairly standard, it is
still not part of the *public* API of this module: you only use the module's
vendored catalog when you have no workspace which could otherwise provide the
same information.  (Yes, workspaces are "coming soon", so we do always need
the vendored catalog info at present; it's on the roadmap to shift this weight
soon, however.)

Why are we also backpedaling and supporting a .timeless/module.tl file?
It's not just for the love of back-compat.
Because if you really want to put everything in one dir (say, for .gitignore
purposes!), then you should be free to do so.
In general, stellar should be as accomodating as possible of whatever directory
structure you want to arrange your projects and source control repositories in.
Want a single folder with a flat set of dirs which all contain one git repo...
and some of them are projects with other authors who haven't merged a module.tl
file yet, so you want to keep working but keep all the timeless config in a
gitignore?  Supported.
But prefer putting the module.tl file in the project root.
It's more obvious there.

Signed-off-by: Eric Myhre <hash@exultant.us>