Hitch supports catalog overlays; test workspace.
Workspace catalog references work.>!!><!<!!! FIREWORKS EVERYWHERE

The catalog name is current hardcoded.  As described in the previous commit's
message, the intent here is to get to supporting overlays of multiple catalogs
(and we do now already have that support in the hitch view interface!), but
I'm taking some shortcuts here to reach demo'ability/PoC.

Amusingly, being overly eager to shove a workspace catalog path into the stack
used by the hitch view interface happens to be fine; if it doesn't exist,
that's still an overall NotFound error which is... correct.

The big remaining hack in the room is warehouse addresses for *outputs*.
I'm making this test pass via a dirty symlink, and this is of course entirely
bonkers and can't last.  Unfortunately, this is less of a hack and more of a
re-discovery of an already bad but-surprisingly-surivable-thus-now-ingrained
hack: there are several places where the ".timeless/warehouse" relative paths
are ducttaped into place, and there's a few API questions to wrangle to fix it.

Signed-off-by: Eric Myhre <hash@exultant.us>