Warehouses for intermediates now in workspaces.
This should perhaps still be even more configurable in the future (read the
workspace.tl file and allow it to configure things?  etc), but solves the
issue of needing per-module hacks to get a bunch of modules in a workspace
to play nice today.

I think workspaces are now... roughly PoC'able?  I'm gonna start trying it.

Signed-off-by: Eric Myhre <hash@exultant.us>