Refactor packages into groups by what duration of activity they are.
This is borrowing some organizational ideas from the Raceway project that seemed to turn out well.

I already fixed one piece of overly application-y code that was in the library ("gadget") area.
There are many more.  In particular the entire CI command.  Which is an actor.  In the main method.  Lul.

Signed-off-by: Eric Myhre <hash@exultant.us>