Maestro implementation draft.
This includes a lot of meta-commentary about the supervisor library,
because I'm co-developing that and this is a proving ground for those
APIs.  So far, it works, though I have Questions about some things.
I'm strongly tempted to make another iteration on the supervision
APIs that *don't* use channels for submission, because this kind of
controller/pool-pair design seems likely to be common, and it already
*has* other channels exposed that have more locally relevant types.
And a bunch of other reasons as well, but I'll save that for the next
design doc in go-sup.

It's also clear from this that we need some more clear type definitions
in go-timeless-api for what module evaluation should *return*.  But
we knew that.

And we can't usefully rig this to anything (namely, replacing the
current hacking CI command) until we also get a watcher actor that we
can pair with this.  That's more than one commit's worth of work, so.

Progress, anyway.  I'm probably going to merge this now, because I
have too many open tasks right now, and I'm not sure the concurrency
engine design stuff is the one I want to be focusing on this week.

Signed-off-by: Eric Myhre <hash@exultant.us>