Jump forward about 18 commits in go-timeless-api.
We've now got Rio and Repeatr *also* all both on this tip of the
"redux" branch, so, for the first time, finally all three of us are
actually building from the same API code, rather than structurally
similar but distinctly maintained branches...!   WHEW!

The repeatr.RunFunc signiture changed back to being significantly
closer to its pre-redux definition.  It turned out there were good
reasons for that.  We now have a repeatr.RunOperation decorator that
handles all of the conversions that Stellar wants to apply.

This unblocks us from making tons more stuff.  For starters, it's now
easy-peasy to call Rio as an RPC from up here in Stellar (as it should
be!), and that's going to let us do excellent stuff like make new
ingest systems which are powered by simply calling `rio pack`.  Soon!

Signed-off-by: Eric Myhre <hash@exultant.us>