Implement "ingest:pack".
This was basically trivial to do after all the groundwork was laid.

(Of course, getting all the API stuff to the point that it becomes
"trivial" to use it like this took a week worth of grinding through
updates across the various repos... but eh, SMOP.)

This should be a huge usability boost.  One very common usecase will
likely be using this feature to ingest a large script file instead of
having an overly long string in action.exec; or to add a bunch of
"patchfiles".  You could do all this with "ingest:git" of course; but
now we've got this much more flexible option as well.

Signed-off-by: Eric Myhre <hash@exultant.us>