Catalog linter: more sanity checks; fix some skip.
The linter should print more warnings now if you somehow aim it at a
non-directory.

And it won't silently skip everything in the catalog if there's a file
with a dot-prefixed name in the root dir.  Yikes.
Turns out returning "SkipDir" when you're on a file in filepath.Walk
does... not do a calm thing.
Discovered that ".git" as a *file* -- as submodules often do -- does
this, so I'd silently not been linting much for a while.  Sigh.

Signed-off-by: Eric Myhre <hash@exultant.us>