Add module; build stellar, tests CLI.
This doesn't run the full tests of stellar because we'd hit the
"nested containers" problem (generally not well solved for any project
the container multiverse right now), and because we'd need to clean
up our dependency management in this repo quite a bit before that would
be feasible anyway.

Actually... I guess that last part would be pretty easy.  We could just
refer back to release builds from the timeless omnibus modules.

But that's not something I'm really going to hammer out impulsively;
there would probably be cleaner more sustainable ways to organize that
and it's not something I want to go depth-first on tonight.

Anyway.  This gives us a nice executable example of our...self.

There's no catalog link included right now because on my computer, this
is in a "workspace" dir, so it's getting the references from there.
We want to introduce tooling which would also replicate the relevant
subset of the catalog into the git repo so everything continues to work
without that external reference being provided, but that's still in the
roadmap rather than really to roll right now.

Signed-off-by: Eric Myhre <hash@exultant.us>