Prepare some helper funcs for isolating test cwd.
Should've done this ages ago; time to scratch it off the list.

Pulled in a rio link to do it.  Could've certainly generated this code
again without pulling in a dependency; but seems unproductive to bother
when we've already got all that lovely logic for handling symlinks and
everything else under the sun.

Signed-off-by: Eric Myhre <hash@exultant.us>