Introduce catalog.SaveCandidateRelease and sagas.
`stellar emerge` will now -- if you have a workspace and we can thus
detect a module name for what we're evaluating -- save the exported
results to a 'catalog.tl' file as a release called "candidate", and
it will do this in a new catalog tree called ".timeless/candidates".

For now, you can copy and paste these over to your real catalog repo
and commit them somewhat manually; in the future, we'll introduce CLI
tools for that as well.

We'll also want to support using these releases as an overlay to a
catalog, so we can do a series of module evaluations at once using
each others "candidate" releases -- and future tooling can then do
releases of *all* of those at once.  The idea is that this sort of
workflow will make it possible to test downstreams before releasing;
one of the oft-evasive but nigh-holy grails of development processes.

Some stubbed methods also suggest we'll soon start writing the module
itself out again as a replay record; this will come after we write up
the module-to-replay reduction and sanity checker logics.

Signed-off-by: Eric Myhre <hash@exultant.us>