Workspace rework!
A series of changes to make things simpler and unify several flows
that had become divergent in the previous draft of workspaces:

- A workspace is now **required** for a module to be evaluable.

- To find a Workspace, we iterate up dirs until finding `.timeless`.
(This is similar to before, but now is even more unconditional.)

- The workspace closest to the cwd will always be preferred.
Combined with the ability to specify a module to evaluate by path,
this means it's possible to evaluate a module at great depth using
the cwd workspace, even if there are other workspace in between.
This is critical to our future stories for multi-project coordination,
while simultaneously being able to maintain full local catalog
vendoring in a workspace per project.

- Modules are now always marked by the 'module.tl' file.
Support for keeping this file inside the '.timeless' dir as a special
case has been dropped; there's just not enough call for it.

- The idea of multiple catalogs per workspace has been *dropped*.
It was a bad idea full of complexity and little utility.
The catalog is now *always* at `.timeless/catalog`.
This fixes a massive UX surprise wherein previously you might have
catalogs at that path (for anon modules with no workspace) or instead
in `.timeless/catalogs/upstream` (note the 's'!); nobody liked that,
and every single person who discovered it was negatively surprised.
So, now we have simplified that problem out of existence.

In smaller details of implementation:

- PathInsideWorkspace was never actually used (and indeed, good lord,
how would have ever have made sense?); dropped it.

- layout.Workspace and layout.Module now both exist separately.
(It's planned to be possible to have one process handle multiple
modules; so we should be able to look up workspace only once.)

- I waffled for a while about whether it's a good idea to let
EvalModule require a workspace.Workspace (i.e., config fully loaded),
because if it's an anonymous module or no candidate is to be saved,
then we don't really *need* it per se.  But, I think worrying about
making it possible for anonymous modules to be executable even if the
workspace config file raises parse errors is kind of... well, silly.
If there are errors, let them error; not conditionally error.

Many test fixtures needed moderate rearrangement.
(Quite a few used the '.timeless/module.tl' file placement, which we
dropped support of... since these tests were approximately the only
actual user of that path.)

Oh, a 'gadget/workspace' package has been introduced.  This contains
the abstraction for mapping module paths on the filesystem into
module names (currently *still* a quite rigid implementation), and
is the future home for any more workspace configuration logic.

I think that's about it.  Whew.

Signed-off-by: Eric Myhre <hash@exultant.us>