Update names of module imports.
These are the names currently used in the content in
https://github.com/polydawn/catalog.git .

While in the long run we intend for *some* names in these positions
to become final, we haven't reached that stage of development yet --
we're still treating all these references as very much "pre 1.0" and
subject to change.

Remember that we also intend to ship tooling for vendoring of catalogs
into other merkle-tracking systems like git, which would make this
change clearer and rather than related to out-of-repo side-effects;
but those are still future work.

Signed-off-by: Eric Myhre <hash@exultant.us>