Resurrect some simple io wrapping libs.
This is code from older versions of the repeatr project.

Probably we'd have better performance if we made a more composable set
of formatting APIs, but this is, well, easier and quicker today.

(Had to update the test system from go-convey to new go-stdlib-t.Run,
and then go-wish for the assertions, but that was about it.)

Gonna use this in a second to prettify some output.

Signed-off-by: Eric Myhre <hash@exultant.us>