Introduce operation.Resolve and operation.Eval.
These gadgets are derived from what was previously a single function,
repeatr.RunOperation, which was in the go-timeless-api repo.  We're
moving it here (and the API repo bump removes it there), and splitting
it apart at the same time.

The newly split functions, and the PreparedOperation struct (or 'prop',
for short) that carries information between them, offer better
opportunities for logging and debugging the resolved states.

Such better logging of the resolved states isn't included in this diff,
but can come up next; this change just gets things over here and split
apart to the degree we can do something.

Signed-off-by: Eric Myhre <hash@exultant.us>