add layout.IsModuleName utility.
This is a step towards addressing some of the TODOs already cropping up
in various CLI commands about the distinction between paths versus
moduleNames.

Worth briefly mentioning: I first tried to make a function of the form
`ParseNameOrPath(lws layout.Workspace, str string) (*api.ModuleName, *layout.Module)`,
but this turned out to be a very twisty and confusing thing, where the
greater part of the documentation turned out to be commenting on what
the function would and would not do based on the fact it doesn't fully
load and understand the workspace config... so, rather than have a
function which has more cautionary codes *than code*, it seems better
to just not attempt to handle that part.  (Maybe more helpful functions
will be useful, but this wasn't the one.)

Signed-off-by: Eric Myhre <hash@exultant.us>