Skippable integration tests; and, module test.
`reach emerge` on this project itself will now run more tests in total
since it can run everything *but* the integration tests.

The integration tests are nontrivial to run because they
A) require more of the other tools on the path (fixable, certainly);
B) nested containers are problematic (less trivial to fix).

Signed-off-by: Eric Myhre <hash@exultant.us>