Build a CLI around this new workspace search system.  Test through it.
I'm thinking that exporting "plumbing" commands will be useful for
letting users poke at what would otherwise be somewhat inscrutable
internal operations.  Good for debugging; good for user understanding;
good for learning; just generally good.  Bonus: it's also awfully
easy to test.

Some parts of this CLI library still feel like they're fighting me a
bit (the error handling is... something; the libraries opinions about
stdout vs stderr are what I'd personally call "wrong"; the structs
are not particularly well documented in code and I have go look at the
website and readmes instead to find even basic info; some of the basic
usage text stuff has very strange opinions; a few other smol details
that are just frustrating... and also a few things that are bigger:
namely, you can't use zero values for a couple of things because they
get overriden by "defaults" and there's not a thing you can do about it
and this is a design flaw that requires a library rewrite to add more
types for maybes in order to address clearly; hej, golang).  But...
it works.

I'm doing granular CLI commands, as well.  Previously in this repo,
I've done all the CLI stuff in the main (or main-adjacent) packages,
and not let it touch anything else.
Here, I'm trying something different.  Getting all the CLI stuff
hoarded in one place made it harder to test, and harder to organize;
let's break it up, and keep it closer to the code it will be calling.
The downside will be that a *lot* of packages will probably end up
importing the CLI package... and if you wanted to use any of these
packages as a library, getting a transitive dep on a CLI package
would certainly seem odd.  So... maybe we'll keep a wee gap there.

Trying a new trick of passing the working direction info around as
if it was a flag.  Seems promising.  Should reduce boilerplate a tiny
bit; also, being able to override it seems a neat bonus.

And the tests!  I wrote some new "shop-made jigs" for this, but they
seem to be turning out well:

- we make a small dummy CLI app, and exercise that...

- ...by calling into it; we don't compile and run the whole thing.
  (I actually tried doing it with a whole "go run" subprocess; that
  also worked, and would let us talk about exit codes more confidently,
  but on the other hand, required a lot more rigamarole (namely, adding
  yet another package, just so it can be called "main", which would
  bring the package count up to *three* per feature, which... no.))

- a "RunIntoBuffer" function just emits strings, and we do a wish
  against them; this works well and simply.

- I made a new "WithTmpdir" function!  There have been several of these
  before (tempdirs for testing come up *many* times in Timeless Stack;
  go figure, right?).  This one is different: it takes the *testing.T
  object as a parameter, and generates the "temp" dir _based on the
  test name_.  It's _not random_.  This is nice because it means that
  it's very easy to write tests against output that contains paths.
  (This is an issue we've previously wrangled with regexp'ing to build
  "normalizers" for output strings (e.g. repeatr tests do some of this)
  but those are kind of complex and kind of a pain; this seems nicer.)
  At present, this won't adapt well to anyone who doesn't have a "/tmp"
  directory... but that's, uh, pretty common; so if we need to flex
  around that someday, well, let's cross that bridge when we come to it.

- There's a wee helper method called "bangFile".  It's really simple;
  all it does is make files or directories, and huck content in them.
  It's meant to be used to set up test fixture data... specifically,
  it's meant to "look good" (align horizontally reasonably well) when
  used in line after line after line in your test code.  The paths
  in particular are supposed to align, to minimize the chance of small
  typos there.  It's *much* simpler than the nearest comparable
  test setup helper functions in Rio... because it does less; we get
  to just completely shrug at permissions, interesting file types,
  mtimes (good lord), etc, here, because none of those things matter
  in this context; and so we do, and its easier to use as a result.
  (I will confess, I also did some probing at creating (yet another)
  filesystem abstraction library recently, because this is definitely
  one of those "recurring friction" things and I don't think any
  gopher has solved it perfectly yet.  But, that's also a big dream
  to chase when you engage with all of it at once, so, here's a
  scope-limited alternative that's quicker to wrap up and ship.)

And that's "it".  This was a fun commit!